backend/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0

# Agent Orchestration
langgraph>=0.2.0
langchain-anthropic>=0.2.0
langchain-core>=0.3.0

# Vector DB
chromadb>=0.5.0

# PDF generation
weasyprint>=62.0

# HTTP client & Voice (Spitch API)
httpx>=0.27.0
jinja2>=3.1.0  # For PDF templates
markdown>=3.7.0
